test-requirements.txt
nose==1.3.7
vcrpy==1.10.3
